Add access-linked environments (#9)
Adds an `Environment` (a.k.a. frame) for tracking name bindings. It currently does not handle type bindings, only variable and function names, and uses access links (or static links) to search through a stack of `EnvironmentLevel`s for a binding of a name. As a result of this change, AST nodes such as `for`, `let`, `while`, `if`, and `assign` are implemented and the integration tests pass. Commits include:

* Add initial function evaluation
* Add native function evaluation
* Fix: 2/1 should evaluate to 2
* Add sequence evaluation, fix integration testing
* Add initial let evaluation
* Make print-tests pass
* Ensure function calls clean up environment
* Add access-linked environments
* Finish for-loop evaluation
* Add while, if, let, assign evaluation